+++ src/components/Stars.jsx
@@ -1,4 +1,6 @@
+import { useState, useEffect, useRef } from 'react'
 import styled from 'styled-components'
+import { motion as m } from 'framer-motion'
 
 const stars1 = `rgb(255, 255, 255) 515px 469px, rgb(255, 255, 255) 253px 1553px, rgb(255, 255, 255) 318px 1209px, rgb(255, 255, 255) 990px 1697px, rgb(255, 255, 255) 714px 1995px, rgb(255, 255, 255) 218px 1305px, rgb(255, 255, 255) 1978px 346px, rgb(255, 255, 255) 1666px 1538px, rgb(255, 255, 255) 54px 1760px, rgb(255, 255, 255) 396px 1163px, rgb(255, 255, 255) 8px 1796px, rgb(255, 255, 255) 485px 185px, rgb(255, 255, 255) 961px 889px, rgb(255, 255, 255) 11px 864px, rgb(255, 255, 255) 600px 1495px, rgb(255, 255, 255) 1538px 620px, rgb(255, 255, 255) 1892px 1493px, rgb(255, 255, 255) 564px 839px, rgb(255, 255, 255) 1870px 1616px, rgb(255, 255, 255) 883px 30px, rgb(255, 255, 255) 1437px 1187px, rgb(255, 255, 255) 1230px 854px, rgb(255, 255, 255) 1070px 273px, rgb(255, 255, 255) 600px 529px, rgb(255, 255, 255) 1215px 1029px, rgb(255, 255, 255) 1385px 854px, rgb(255, 255, 255) 386px 1618px, rgb(255, 255, 255) 1994px 957px, rgb(255, 255, 255) 1782px 489px, rgb(255, 255, 255) 1398px 1863px, rgb(255, 255, 255) 572px 1631px, rgb(255, 255, 255) 1658px 266px, rgb(255, 255, 255) 599px 1706px, rgb(255, 255, 255) 1102px 928px, rgb(255, 255, 255) 159px 1165px, rgb(255, 255, 255) 81px 1296px, rgb(255, 255, 255) 467px 112px, rgb(255, 255, 255) 1059px 1649px, rgb(255, 255, 255) 1476px 439px, rgb(255, 255, 255) 1183px 390px, rgb(255, 255, 255) 1270px 606px, rgb(255, 255, 255) 878px 1886px, rgb(255, 255, 255) 828px 868px, rgb(255, 255, 255) 464px 1310px, rgb(255, 255, 255) 1849px 1050px, rgb(255, 255, 255) 1321px 1480px, rgb(255, 255, 255) 1172px 1929px, rgb(255, 255, 255) 1763px 1836px, rgb(255, 255, 255) 1126px 1903px, rgb(255, 255, 255) 973px 1604px, rgb(255, 255, 255) 1807px 66px, rgb(255, 255, 255) 674px 59px, rgb(255, 255, 255) 412px 1015px, rgb(255, 255, 255) 543px 707px, rgb(255, 255, 255) 495px 1186px, rgb(255, 255, 255) 123px 544px, rgb(255, 255, 255) 416px 15px, rgb(255, 255, 255) 1711px 651px, rgb(255, 255, 255) 265px 1407px, rgb(255, 255, 255) 1206px 1766px, rgb(255, 255, 255) 141px 311px, rgb(255, 255, 255) 1238px 1995px, rgb(255, 255, 255) 1582px 1419px, rgb(255, 255, 255) 803px 1520px, rgb(255, 255, 255) 520px 1884px, rgb(255, 255, 255) 1534px 142px, rgb(255, 255, 255) 1744px 283px, rgb(255, 255, 255) 546px 1950px, rgb(255, 255, 255) 1185px 303px, rgb(255, 255, 255) 1102px 296px, rgb(255, 255, 255) 1189px 1247px, rgb(255, 255, 255) 1497px 863px, rgb(255, 255, 255) 962px 1646px, rgb(255, 255, 255) 979px 1841px, rgb(255, 255, 255) 1028px 1409px, rgb(255, 255, 255) 386px 1253px, rgb(255, 255, 255) 1032px 474px, rgb(255, 255, 255) 1687px 452px, rgb(255, 255, 255) 1801px 1401px, rgb(255, 255, 255) 528px 740px, rgb(255, 255, 255) 455px 385px, rgb(255, 255, 255) 299px 1323px, rgb(255, 255, 255) 585px 1102px, rgb(255, 255, 255) 1850px 679px, rgb(255, 255, 255) 318px 253px, rgb(255, 255, 255) 1493px 1161px, rgb(255, 255, 255) 775px 1392px, rgb(255, 255, 255) 968px 296px, rgb(255, 255, 255) 1426px 145px, rgb(255, 255, 255) 1969px 169px, rgb(255, 255, 255) 157px 1755px, rgb(255, 255, 255) 64px 1531px, rgb(255, 255, 255) 165px 1255px, rgb(255, 255, 255) 828px 714px, rgb(255, 255, 255) 562px 1501px, rgb(255, 255, 255) 1186px 1466px, rgb(255, 255, 255) 1216px 695px, rgb(255, 255, 255) 220px 1081px, rgb(255, 255, 255) 1583px 1347px, rgb(255, 255, 255) 547px 909px, rgb(255, 255, 255) 405px 1946px, rgb(255, 255, 255) 933px 1297px, rgb(255, 255, 255) 1384px 455px, rgb(255, 255, 255) 328px 1142px, rgb(255, 255, 255) 1031px 1142px, rgb(255, 255, 255) 1003px 813px, rgb(255, 255, 255) 632px 766px, rgb(255, 255, 255) 643px 1545px, rgb(255, 255, 255) 463px 816px, rgb(255, 255, 255) 415px 667px, rgb(255, 255, 255) 465px 1212px, rgb(255, 255, 255) 72px 116px, rgb(255, 255, 255) 1883px 1493px, rgb(255, 255, 255) 399px 404px, rgb(255, 255, 255) 1876px 23px, rgb(255, 255, 255) 1889px 1486px, rgb(255, 255, 255) 1309px 1174px, rgb(255, 255, 255) 984px 1457px, rgb(255, 255, 255) 1998px 1939px, rgb(255, 255, 255) 1017px 657px, rgb(255, 255, 255) 856px 1212px, rgb(255, 255, 255) 1571px 132px, rgb(255, 255, 255) 1009px 355px, rgb(255, 255, 255) 1710px 3px, rgb(255, 255, 255) 1971px 170px, rgb(255, 255, 255) 505px 297px, rgb(255, 255, 255) 714px 224px, rgb(255, 255, 255) 1352px 1480px, rgb(255, 255, 255) 296px 1845px, rgb(255, 255, 255) 1638px 209px, rgb(255, 255, 255) 1408px 335px, rgb(255, 255, 255) 1566px 609px, rgb(255, 255, 255) 113px 922px, rgb(255, 255, 255) 344px 1219px, rgb(255, 255, 255) 711px 966px, rgb(255, 255, 255) 632px 1380px, rgb(255, 255, 255) 1503px 1899px, rgb(255, 255, 255) 250px 1979px, rgb(255, 255, 255) 905px 1258px, rgb(255, 255, 255) 1422px 13px, rgb(255, 255, 255) 942px 1px, rgb(255, 255, 255) 1777px 1691px, rgb(255, 255, 255) 1492px 783px, rgb(255, 255, 255) 459px 1889px, rgb(255, 255, 255) 1660px 1274px, rgb(255, 255, 255) 652px 745px, rgb(255, 255, 255) 1754px 187px, rgb(255, 255, 255) 1298px 583px, rgb(255, 255, 255) 987px 1919px, rgb(255, 255, 255) 852px 1914px, rgb(255, 255, 255) 798px 1519px, rgb(255, 255, 255) 972px 1921px, rgb(255, 255, 255) 1725px 1449px, rgb(255, 255, 255) 822px 1646px, rgb(255, 255, 255) 1992px 1758px, rgb(255, 255, 255) 1963px 838px, rgb(255, 255, 255) 733px 102px, rgb(255, 255, 255) 906px 1231px, rgb(255, 255, 255) 354px 785px, rgb(255, 255, 255) 1075px 1315px, rgb(255, 255, 255) 1255px 677px, rgb(255, 255, 255) 1476px 252px, rgb(255, 255, 255) 1759px 669px, rgb(255, 255, 255) 1409px 808px, rgb(255, 255, 255) 827px 962px, rgb(255, 255, 255) 624px 552px, rgb(255, 255, 255) 1074px 1577px, rgb(255, 255, 255) 1634px 250px, rgb(255, 255, 255) 859px 1515px, rgb(255, 255, 255) 955px 1907px, rgb(255, 255, 255) 633px 1137px, rgb(255, 255, 255) 718px 1677px, rgb(255, 255, 255) 1581px 431px, rgb(255, 255, 255) 1857px 1020px, rgb(255, 255, 255) 782px 1098px, rgb(255, 255, 255) 832px 78px, rgb(255, 255, 255) 1233px 1326px, rgb(255, 255, 255) 945px 1800px, rgb(255, 255, 255) 305px 705px, rgb(255, 255, 255) 443px 540px, rgb(255, 255, 255) 1704px 972px, rgb(255, 255, 255) 1156px 1309px, rgb(255, 255, 255) 89px 65px, rgb(255, 255, 255) 549px 1127px, rgb(255, 255, 255) 1726px 1649px, rgb(255, 255, 255) 1624px 78px, rgb(255, 255, 255) 1331px 1280px, rgb(255, 255, 255) 1783px 1916px, rgb(255, 255, 255) 1892px 1440px, rgb(255, 255, 255) 197px 1456px, rgb(255, 255, 255) 337px 1252px, rgb(255, 255, 255) 1637px 1159px, rgb(255, 255, 255) 1209px 410px, rgb(255, 255, 255) 1130px 1437px, rgb(255, 255, 255) 549px 1650px, rgb(255, 255, 255) 1173px 1687px, rgb(255, 255, 255) 801px 224px, rgb(255, 255, 255) 1213px 1586px, rgb(255, 255, 255) 1610px 1152px, rgb(255, 255, 255) 1916px 1756px, rgb(255, 255, 255) 271px 1261px, rgb(255, 255, 255) 707px 376px, rgb(255, 255, 255) 438px 1343px, rgb(255, 255, 255) 1773px 522px, rgb(255, 255, 255) 1779px 1546px, rgb(255, 255, 255) 1989px 1133px, rgb(255, 255, 255) 133px 345px, rgb(255, 255, 255) 1533px 1364px, rgb(255, 255, 255) 918px 154px, rgb(255, 255, 255) 1379px 1078px, rgb(255, 255, 255) 1734px 303px, rgb(255, 255, 255) 1843px 275px, rgb(255, 255, 255) 1257px 388px, rgb(255, 255, 255) 1758px 1175px, rgb(255, 255, 255) 1556px 967px, rgb(255, 255, 255) 672px 1585px, rgb(255, 255, 255) 491px 1317px, rgb(255, 255, 255) 1000px 1686px, rgb(255, 255, 255) 1236px 93px, rgb(255, 255, 255) 671px 40px, rgb(255, 255, 255) 1136px 1542px, rgb(255, 255, 255) 968px 512px, rgb(255, 255, 255) 1191px 1970px, rgb(255, 255, 255) 1907px 1428px, rgb(255, 255, 255) 1630px 902px, rgb(255, 255, 255) 1296px 1539px, rgb(255, 255, 255) 374px 819px, rgb(255, 255, 255) 709px 126px, rgb(255, 255, 255) 497px 1047px, rgb(255, 255, 255) 990px 620px, rgb(255, 255, 255) 489px 1929px, rgb(255, 255, 255) 797px 33px, rgb(255, 255, 255) 1306px 562px, rgb(255, 255, 255) 196px 1278px, rgb(255, 255, 255) 828px 1419px, rgb(255, 255, 255) 1895px 993px, rgb(255, 255, 255) 1241px 1101px, rgb(255, 255, 255) 1626px 1157px, rgb(255, 255, 255) 1329px 311px, rgb(255, 255, 255) 1653px 356px, rgb(255, 255, 255) 1733px 596px, rgb(255, 255, 255) 1062px 732px, rgb(255, 255, 255) 1013px 574px, rgb(255, 255, 255) 59px 1619px, rgb(255, 255, 255) 354px 597px, rgb(255, 255, 255) 1679px 984px, rgb(255, 255, 255) 1375px 1413px, rgb(255, 255, 255) 197px 1177px, rgb(255, 255, 255) 578px 1513px, rgb(255, 255, 255) 446px 172px, rgb(255, 255, 255) 1585px 1578px, rgb(255, 255, 255) 395px 1328px, rgb(255, 255, 255) 1643px 1470px, rgb(255, 255, 255) 432px 1954px, rgb(255, 255, 255) 578px 348px, rgb(255, 255, 255) 832px 1060px, rgb(255, 255, 255) 1993px 1088px, rgb(255, 255, 255) 1152px 176px, rgb(255, 255, 255) 418px 795px, rgb(255, 255, 255) 1813px 27px, rgb(255, 255, 255) 1617px 599px, rgb(255, 255, 255) 357px 1115px, rgb(255, 255, 255) 1791px 1534px, rgb(255, 255, 255) 961px 1312px, rgb(255, 255, 255) 85px 1920px, rgb(255, 255, 255) 289px 1714px, rgb(255, 255, 255) 838px 1888px, rgb(255, 255, 255) 1244px 1061px, rgb(255, 255, 255) 199px 21px, rgb(255, 255, 255) 1789px 162px, rgb(255, 255, 255) 712px 970px, rgb(255, 255, 255) 1492px 45px, rgb(255, 255, 255) 234px 1347px, rgb(255, 255, 255) 960px 1498px, rgb(255, 255, 255) 1355px 1485px, rgb(255, 255, 255) 370px 304px, rgb(255, 255, 255) 1741px 1341px, rgb(255, 255, 255) 629px 1345px, rgb(255, 255, 255) 356px 1711px, rgb(255, 255, 255) 1868px 277px, rgb(255, 255, 255) 336px 900px, rgb(255, 255, 255) 96px 1256px, rgb(255, 255, 255) 653px 989px, rgb(255, 255, 255) 1810px 1344px, rgb(255, 255, 255) 1468px 790px, rgb(255, 255, 255) 1417px 1806px, rgb(255, 255, 255) 490px 417px, rgb(255, 255, 255) 9px 189px, rgb(255, 255, 255) 569px 1465px, rgb(255, 255, 255) 338px 371px, rgb(255, 255, 255) 916px 803px, rgb(255, 255, 255) 326px 288px, rgb(255, 255, 255) 32px 749px, rgb(255, 255, 255) 145px 847px, rgb(255, 255, 255) 999px 708px, rgb(255, 255, 255) 434px 1px, rgb(255, 255, 255) 1532px 1733px, rgb(255, 255, 255) 1499px 1885px, rgb(255, 255, 255) 137px 1658px, rgb(255, 255, 255) 671px 353px, rgb(255, 255, 255) 1580px 1175px, rgb(255, 255, 255) 890px 779px, rgb(255, 255, 255) 151px 641px, rgb(255, 255, 255) 1579px 356px, rgb(255, 255, 255) 1522px 1645px, rgb(255, 255, 255) 80px 25px, rgb(255, 255, 255) 817px 391px, rgb(255, 255, 255) 1161px 191px, rgb(255, 255, 255) 875px 1453px, rgb(255, 255, 255) 1922px 457px, rgb(255, 255, 255) 963px 1714px, rgb(255, 255, 255) 1644px 531px, rgb(255, 255, 255) 1820px 1904px, rgb(255, 255, 255) 444px 1746px, rgb(255, 255, 255) 1185px 1237px, rgb(255, 255, 255) 1972px 822px, rgb(255, 255, 255) 1054px 455px, rgb(255, 255, 255) 356px 1723px, rgb(255, 255, 255) 1367px 1361px, rgb(255, 255, 255) 1813px 345px, rgb(255, 255, 255) 1654px 717px, rgb(255, 255, 255) 1599px 885px, rgb(255, 255, 255) 1796px 1223px, rgb(255, 255, 255) 1659px 701px, rgb(255, 255, 255) 598px 1833px, rgb(255, 255, 255) 1841px 513px, rgb(255, 255, 255) 1041px 645px, rgb(255, 255, 255) 513px 1712px, rgb(255, 255, 255) 1478px 661px, rgb(255, 255, 255) 1925px 1361px, rgb(255, 255, 255) 475px 1559px, rgb(255, 255, 255) 1615px 1304px, rgb(255, 255, 255) 818px 223px, rgb(255, 255, 255) 968px 830px, rgb(255, 255, 255) 220px 343px, rgb(255, 255, 255) 545px 1767px, rgb(255, 255, 255) 704px 299px, rgb(255, 255, 255) 528px 1628px, rgb(255, 255, 255) 1755px 1486px, rgb(255, 255, 255) 1851px 1852px, rgb(255, 255, 255) 631px 1740px, rgb(255, 255, 255) 1461px 507px, rgb(255, 255, 255) 1101px 1369px, rgb(255, 255, 255) 1797px 270px, rgb(255, 255, 255) 196px 406px, rgb(255, 255, 255) 959px 1517px, rgb(255, 255, 255) 798px 250px, rgb(255, 255, 255) 1745px 1301px, rgb(255, 255, 255) 1302px 1605px, rgb(255, 255, 255) 1581px 132px, rgb(255, 255, 255) 1370px 423px, rgb(255, 255, 255) 675px 1549px, rgb(255, 255, 255) 179px 752px, rgb(255, 255, 255) 1821px 1036px, rgb(255, 255, 255) 875px 911px, rgb(255, 255, 255) 1396px 1314px, rgb(255, 255, 255) 576px 440px, rgb(255, 255, 255) 969px 1165px, rgb(255, 255, 255) 1026px 703px, rgb(255, 255, 255) 555px 107px, rgb(255, 255, 255) 173px 365px, rgb(255, 255, 255) 596px 1542px, rgb(255, 255, 255) 1099px 1421px, rgb(255, 255, 255) 42px 748px, rgb(255, 255, 255) 1953px 1722px, rgb(255, 255, 255) 1404px 1816px, rgb(255, 255, 255) 60px 71px, rgb(255, 255, 255) 897px 804px, rgb(255, 255, 255) 1356px 86px, rgb(255, 255, 255) 1567px 1816px, rgb(255, 255, 255) 1061px 1850px, rgb(255, 255, 255) 1292px 504px, rgb(255, 255, 255) 1288px 1173px, rgb(255, 255, 255) 1732px 1406px, rgb(255, 255, 255) 661px 1459px, rgb(255, 255, 255) 1071px 314px, rgb(255, 255, 255) 512px 118px, rgb(255, 255, 255) 1122px 1550px, rgb(255, 255, 255) 1888px 1764px, rgb(255, 255, 255) 160px 463px, rgb(255, 255, 255) 676px 299px, rgb(255, 255, 255) 1408px 919px, rgb(255, 255, 255) 1167px 207px, rgb(255, 255, 255) 1886px 1824px, rgb(255, 255, 255) 94px 395px, rgb(255, 255, 255) 641px 844px, rgb(255, 255, 255) 1183px 1155px, rgb(255, 255, 255) 180px 638px, rgb(255, 255, 255) 83px 1677px, rgb(255, 255, 255) 90px 572px, rgb(255, 255, 255) 61px 768px, rgb(255, 255, 255) 236px 1921px, rgb(255, 255, 255) 689px 1404px, rgb(255, 255, 255) 157px 1986px, rgb(255, 255, 255) 885px 681px, rgb(255, 255, 255) 1270px 1070px, rgb(255, 255, 255) 617px 1837px, rgb(255, 255, 255) 1908px 1693px, rgb(255, 255, 255) 1404px 1300px, rgb(255, 255, 255) 871px 839px`
 
@@ -6,22 +8,19 @@ const stars2 = `rgb(255, 255, 255) 1091px 1534px, rgb(255, 255, 255) 1063px 836p
 
 const stars3 = `rgb(255, 255, 255) 544px 1474px, rgb(255, 255, 255) 1230px 546px, rgb(255, 255, 255) 567px 1497px, rgb(255, 255, 255) 392px 1831px, rgb(255, 255, 255) 578px 919px, rgb(255, 255, 255) 974px 1532px, rgb(255, 255, 255) 1132px 94px, rgb(255, 255, 255) 955px 783px, rgb(255, 255, 255) 807px 789px, rgb(255, 255, 255) 940px 776px, rgb(255, 255, 255) 1144px 500px, rgb(255, 255, 255) 477px 1535px, rgb(255, 255, 255) 1328px 1429px, rgb(255, 255, 255) 950px 895px, rgb(255, 255, 255) 853px 1271px, rgb(255, 255, 255) 1681px 980px, rgb(255, 255, 255) 1826px 1837px, rgb(255, 255, 255) 1400px 473px, rgb(255, 255, 255) 1621px 715px, rgb(255, 255, 255) 1625px 1277px, rgb(255, 255, 255) 113px 168px, rgb(255, 255, 255) 986px 1569px, rgb(255, 255, 255) 730px 613px, rgb(255, 255, 255) 1645px 1840px, rgb(255, 255, 255) 448px 1769px, rgb(255, 255, 255) 929px 841px, rgb(255, 255, 255) 584px 1943px, rgb(255, 255, 255) 145px 258px, rgb(255, 255, 255) 518px 1750px, rgb(255, 255, 255) 847px 1730px, rgb(255, 255, 255) 721px 271px, rgb(255, 255, 255) 1399px 1653px, rgb(255, 255, 255) 1125px 630px, rgb(255, 255, 255) 740px 516px, rgb(255, 255, 255) 1950px 314px, rgb(255, 255, 255) 1098px 634px, rgb(255, 255, 255) 407px 1399px, rgb(255, 255, 255) 231px 554px, rgb(255, 255, 255) 1881px 1042px, rgb(255, 255, 255) 320px 1733px`
 
-const getRandomInt = (max) => {
-  return Math.floor(Math.random() * max);
-}
+// const getRandomInt = (max) => {
+//   return Math.floor(Math.random() * max);
+// }
 
-const multipleBoxShadow = (n) => {
-  let boxShadow = ''
-  for(let i = 0; i < n; i++){
-      boxShadow += `${getRandomInt(2000)}px ${getRandomInt(2000)}px #FFF, `
-  }
-  return boxShadow.slice(0, -2)
-}
+// const multipleBoxShadow = (n) => {
+//   let boxShadow = ''
+//   for(let i = 0; i < n; i++){
+//       boxShadow += `${getRandomInt(2000)}px ${getRandomInt(2000)}px #FFF, `
+//   }
+//   return boxShadow.slice(0, -2)
+// }
 
 const Span = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
   width: ${p=>p.size || 1}px;
   height: ${p=>p.size || 1}px;
   background: transparent;
@@ -40,10 +39,36 @@ const Span = styled.div`
 `
 
 const Stars = () => {
+  const [point, setPoint] = useState({ x: 0, y: 0 });
+  const { x, y } = point;
+  const ref = useRef();
+
+  useEffect(() => {
+    const handlePointerMove = ({ clientX, clientY }) => {
+      const element = ref.current;
+      const parallaxFactor = 65;
+      const x = (clientX - element.offsetLeft - element.offsetWidth / 2);
+      const y = (clientY - element.offsetTop - element.offsetHeight / 2);
+      setPoint({ x, y});
+    };
+
+    window.addEventListener("mousemove", handlePointerMove);
+
+    return () => {
+      window.removeEventListener("mousemove", handlePointerMove);
+    };
+  }, []);
+  
   return <>
-    <Span stars={stars1} size={1.5} timing={120} />
-    <Span stars={stars2} size={2.4} timing={190}  />
-    <Span stars={stars3} size={3.5} timing={230}  />
+    <m.div style={{ position: 'absolute', top: 0, left: 0, transform: `translate(${x / 50}px, ${y / 50}px)` }}>
+      <Span ref={ref} stars={stars1} size={1.5} timing={120} />
+    </m.div>
+    <m.div style={{ position: 'absolute', top: 0, left: 0, transform: `translate(${x / 85}px, ${y / 85}px)` }}>
+      <Span stars={stars2} size={2.4} timing={190} />
+    </m.div>
+    <m.div style={{ position: 'absolute', top: 0, left: 0, transform: `translate(${x / 120}px, ${y / 120}px)` }}>
+      <Span stars={stars3} size={3.5} timing={230} />
+    </m.div>
   </>
 }
